-Created separate files to organize functions. Make sure to run program in 'SUDO' for full use of program! -Moved test scripts into its own directory.
